LayoutTests/imported/w3c/web-platform-tests/css/css-text/overflow-wrap/overflow-wrap-anywhere-span-002-expected.txt

1234567890

FAIL #container 1 assert_equals: 
<div id="container" data-expected-height="20">12345678<span>90</span></div>
height expected 20 but got 10
